pyproject.toml
[project]
name = "mini-lewm"
version = "0.1.0"
description = "Minimal LeWorldModel-style JEPA on PointWorld"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib",
  "numpy",
  "pydantic",
  "pyyaml",
  "scikit-learn",
  "torch",
  "torchvision",
  "tqdm",
]

[dependency-groups]
dev = [
  "pytest",
  "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mini_lewm"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
